--- src/config/locales/index.js
@@ -1,69 +1,0 @@
-export default {
-  'en-US': require('./en-US.json'),
-  'hi-IN': require('./hi-IN.json'),
-  'id-ID': require('./id-ID.json'),
-  'bn-BD': require('./bn-BD.json'),
-  'de-DE': require('./de-DE.json'),
-  'es-ES': require('./es-ES.json'),
-  'fr-FR': require('./fr-FR.json'),
-  'hu-HU': require('./hu-HU.json'),
-  'bg-BG': require('./bg-BG.json'),
-  'it-IT': require('./it-IT.json'),
-  'ja-JP': require('./ja-JP.json'),
-  'ru-RU': require('./ru-RU.json'),
-  'tr-TR': require('./tr-TR.json'),
-  'ko-KR': require('./ko-KR.json'),
-  'uk-UA': require('./uk-UA.json'),
-  'lt-LT': require('./lt-LT.json'),
-  'fi-FI': require('./fi-FI.json'),
-  'fil-PH': require('./fil-PH.json'),
-  'pt-PT': require('./pt-PT.json'),
-  'pl-PL': require('./pl-PL.json'),
-  'sr-CS': require('./sr-CS.json'),
-  'fa-IR': require('./fa-IR.json'),
-  'ar-SA': require('./ar-SA.json'),
-  'az-AZ': require('./az-AZ.json'),
-  'uz-UZ': require('./uz-UZ.json'),
-  'ro-RO': require('./ro-RO.json'),
-  'ms-MY': require('./ms-MY.json'),
-  'ku-TR': require('./ku-TR.json'),
-  'vi-VN': require('./vi-VN.json'),
-  'zh-CN': require('./zh-CN.json'),
-  'zh-TW': require('./zh-TW.json'),
-  'et-EE': require('./et-EE.json'),
-};
-
-export const locales = [
-  { id: 'en-US', name: 'English' },
-  { id: 'hi-IN', name: 'Hindi' },
-  { id: 'id-ID', name: 'Indonesian' },
-  { id: 'bn-BD', name: 'Bengali' },
-  { id: 'de-DE', name: 'Deutsche' },
-  { id: 'es-ES', name: 'Spanish' },
-  { id: 'fr-FR', name: 'French' },
-  { id: 'hu-HU', name: 'Hungarian' },
-  { id: 'bg-BG', name: 'Bulgarian' },
-  { id: 'it-IT', name: 'Italian' },
-  { id: 'ja-JP', name: 'Japanese' },
-  { id: 'ru-RU', name: 'Русский' },
-  { id: 'tr-TR', name: 'Türkçe' },
-  { id: 'ko-KR', name: 'Korean' },
-  { id: 'uk-UA', name: 'Ukrainian' },
-  { id: 'lt-LT', name: 'Lithuanian' },
-  { id: 'fi-FI', name: 'Finnish' },
-  { id: 'fil-PH', name: 'Filipino' },
-  { id: 'pt-PT', name: 'Porteguese' },
-  { id: 'et-EE', name: 'Estonian' },
-  { id: 'pl-PL', name: 'Polish' },
-  { id: 'sr-CS', name: 'Serbian' },
-  { id: 'fa-IR', name: 'Persian' },
-  { id: 'ar-SA', name: 'Arabic' },
-  { id: 'az-AZ', name: 'Azerbaijani' },
-  { id: 'uz-UZ', name: "O'zbek" },
-  { id: 'ro-RO', name: 'Romanian' },
-  { id: 'ms-MY', name: 'Malay' },
-  { id: 'ku-TR', name: 'Kurdish' },
-  { id: 'vi-VN', name: 'Vietnamese' },
-  { id: 'zh-CN', name: 'Chinese-Simplified' },
-  { id: 'zh-TW', name: 'Chinese-Traditional' },
-];
