+++ src/screens/searchResult/index.ts
